Refactored the `accountNumber` attribute in the Java domain to support big integer
auto-generation and introduced the core DDL scripts under a new directory.

* Domain Refactoring: Changed `accountNumber` from String to Long in the Account model
  to align with the PostgreSQL performance requirements for auto-generated values.
* Database Infrastructure: Created a new `/scripts` directory at the core module root
  containing the initial relational schema definitions (`users`, `accounts`, `clients`).
* Auto-Generation Strategy: Configured `account_number` as `BIGINT GENERATED ALWAYS AS IDENTITY`
  starting at 1,000,000 to prevent data truncation and guarantee safe sequence increments.
* Data Integrity: Enforced strict constraints and explicit restrictive foreign keys,
  avoiding cascade deletions to protect auditing compliance.

+++ core/src/main/java/com/sandboxbank/model/account/Account.java
@@ -14,7 +14,7 @@ import java.util.Locale;
 public class Account {
 
     private Long accountID;
-    private int accountNumber;
+    private Long accountNumber;
     private AccountType type;
     private AccountStatus status;
     private String pinHash;
@@ -35,7 +35,7 @@ public class Account {
     }
 
 
-    public Account(Long accountID, int accountNumber, AccountType type, AccountStatus status, String pinHash, MonetaryAmount balance, LocalDateTime createdAt, LocalDateTime updatedAt) {
+    public Account(Long accountID, Long accountNumber, AccountType type, AccountStatus status, String pinHash, MonetaryAmount balance, LocalDateTime createdAt, LocalDateTime updatedAt) {
 
         this.accountID = accountID;
         this.accountNumber = accountNumber;
@@ -57,12 +57,12 @@ public class Account {
         this.accountID = accountID;
     }
 
-    public int getAccountNumber() {
+    public Long getAccountNumber() {
 
         return accountNumber;
     }
 
-    public void setAccountNumber(int accountNumber) {
+    public void setAccountNumber(Long accountNumber) {
 
         this.accountNumber = accountNumber;
     }
